Revised notification datatypes; moved notifer file to pipeline/notifications; revised notification test cases
