+++ pipeline/notifications/__init__.py
@@ -1,1 +1,2 @@
 from pipeline.notifications.check_message import *
+from pipeline.notifications.notifier import *
